Support log logging exporter via autoconfigure
