+++ sdk-extensions/autoconfigure/src/testOtlpGrpc/java/io/opentelemetry/sdk/autoconfigure/OtlpGrpcConfigTest.java
@@ -79,7 +79,7 @@ class OtlpGrpcConfigTest {
     MetricExporter metricExporter =
         MetricExporterConfiguration.configureOtlpMetrics(properties, SdkMeterProvider.builder());
     LogExporter logExporter =
-        LogEmitterProviderConfiguration.configureOtlpLogs(properties, MeterProvider.noop());
+        LogExporterConfiguration.configureOtlpLogs(properties, MeterProvider.noop());
 
     assertThat(spanExporter)
         .extracting("delegate.timeoutNanos")
@@ -210,7 +210,7 @@ class OtlpGrpcConfigTest {
     props.put("otel.exporter.otlp.logs.compression", "gzip");
     props.put("otel.exporter.otlp.logs.timeout", "15s");
     LogExporter logExporter =
-        LogEmitterProviderConfiguration.configureOtlpLogs(
+        LogExporterConfiguration.configureOtlpLogs(
             DefaultConfigProperties.createForTest(props), MeterProvider.noop());
 
     assertThat(logExporter)
@@ -248,8 +248,7 @@ class OtlpGrpcConfigTest {
         .hasMessageContaining("Invalid OTLP certificate path:");
 
     assertThatThrownBy(
-            () ->
-                LogEmitterProviderConfiguration.configureOtlpLogs(properties, MeterProvider.noop()))
+            () -> LogExporterConfiguration.configureOtlpLogs(properties, MeterProvider.noop()))
         .isInstanceOf(ConfigurationException.class)
         .hasMessageContaining("Invalid OTLP certificate path:");
   }
