+++ sdk-extensions/autoconfigure/src/testOtlpHttp/java/io/opentelemetry/sdk/autoconfigure/OtlpHttpRetryTest.java
@@ -97,7 +97,7 @@ class OtlpHttpRetryTest {
         server.selfSignedCertificate.certificate().getPath());
     props.put("otel.experimental.exporter.otlp.retry.enabled", "true");
     LogExporter logExporter =
-        LogEmitterProviderConfiguration.configureOtlpLogs(
+        LogExporterConfiguration.configureOtlpLogs(
             DefaultConfigProperties.createForTest(props), MeterProvider.noop());
 
     testRetryableStatusCodes(() -> LOG_DATA, logExporter::export, server.logRequests::size);
